MODE INICIAL ACABADO, Salida esperada ya es int

+++ ros2_ws/src/stats/src/mode.cpp
@@ -2,14 +2,16 @@
 #include "std_msgs/msg/float32.hpp"
 #include "std_msgs/msg/int32.hpp"
 #include <map>
+#include <vector>
 #include <algorithm>
 
 
-std::shared_ptr<rclcpp::Publisher<std_msgs::msg::Float32>> publisher; //declaro el subscriptor para el callback
+std::shared_ptr<rclcpp::Publisher<std_msgs::msg::Int32>> publisher; //declaro el subscriptor para el callback
 std::map<float,int> counter;
 int mayor = 0;
 float valor_mayor = 0.0;
 void topic_callback(const std_msgs::msg::Int32::SharedPtr msg){ 
+  
     float num = msg->data;
     counter[num]+=1;
     
@@ -21,7 +23,8 @@ void topic_callback(const std_msgs::msg::Int32::SharedPtr msg){
     
     }
     
-    std_msgs::msg::Float32 out_msg;
+    std_msgs::msg::Int32 out_msg;
+    
     out_msg.data = valor_mayor;
     publisher->publish(out_msg);
 }
@@ -33,7 +36,7 @@ int main(int argc, char * argv[]){
     auto subscription = 
         node->create_subscription<std_msgs::msg::Int32>(
             "number", 10, topic_callback); //me suscribo a topic
-    publisher = node->create_publisher<std_msgs::msg::Float32>("topic_mode", 10); //publico la media
+    publisher = node->create_publisher<std_msgs::msg::Int32>("topic_mode", 10); //publico la media
 
     rclcpp::spin(node); 
     rclcpp::shutdown();
